hotel_api/requirements.txt
Django>=5.0
django-ninja>=1.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
gunicorn>=21.0
uvicorn[standard]>=0.28